fix: create help to get last day month
